one_for_all_func/requirements.txt
bertopic==0.14.1
docx==0.2.4
gensim==4.3.0
matplotlib==3.7.0
nltk==3.7
pandas==1.5.3
Pillow==10.1.0
python_docx==0.8.11
streamlit==1.28.2
